+++ juniper_cascor_client/ws_client.py
@@ -37,7 +37,29 @@ from juniper_cascor_protocol.envelope import UnknownEnvelope, validate_envelope
 from websockets.asyncio.client import ClientConnection
 from websockets.protocol import State
 
-from juniper_cascor_client.constants import API_KEY_ENV_VAR, API_KEY_HEADER_NAME, DEFAULT_CONTROL_STREAM_TIMEOUT, DEFAULT_LIVENESS_WINDOW_SEC, DEFAULT_SET_PARAMS_TIMEOUT, DEFAULT_WS_BASE_URL, MAX_PENDING_COMMANDS, WS_CONTROL_PATH, WS_MSG_TYPE_CANDIDATE_PROGRESS, WS_MSG_TYPE_CASCADE_ADD, WS_MSG_TYPE_COMMAND_OUT, WS_MSG_TYPE_COMMAND_RESPONSE, WS_MSG_TYPE_CONNECTION_ESTABLISHED, WS_MSG_TYPE_EVENT, WS_MSG_TYPE_METRICS, WS_MSG_TYPE_PING, WS_MSG_TYPE_PONG, WS_MSG_TYPE_STATE, WS_MSG_TYPE_TOPOLOGY, WS_ORIGIN_ENV_VAR, WS_TRAINING_PATH
+from juniper_cascor_client.constants import (
+    API_KEY_ENV_VAR,
+    API_KEY_HEADER_NAME,
+    DEFAULT_CONTROL_STREAM_TIMEOUT,
+    DEFAULT_LIVENESS_WINDOW_SEC,
+    DEFAULT_SET_PARAMS_TIMEOUT,
+    DEFAULT_WS_BASE_URL,
+    MAX_PENDING_COMMANDS,
+    WS_CONTROL_PATH,
+    WS_MSG_TYPE_CANDIDATE_PROGRESS,
+    WS_MSG_TYPE_CASCADE_ADD,
+    WS_MSG_TYPE_COMMAND_OUT,
+    WS_MSG_TYPE_COMMAND_RESPONSE,
+    WS_MSG_TYPE_CONNECTION_ESTABLISHED,
+    WS_MSG_TYPE_EVENT,
+    WS_MSG_TYPE_METRICS,
+    WS_MSG_TYPE_PING,
+    WS_MSG_TYPE_PONG,
+    WS_MSG_TYPE_STATE,
+    WS_MSG_TYPE_TOPOLOGY,
+    WS_ORIGIN_ENV_VAR,
+    WS_TRAINING_PATH,
+)
 from juniper_cascor_client.exceptions import JuniperCascorClientError, JuniperCascorConnectionError, JuniperCascorOverloadError, JuniperCascorTimeoutError
 from juniper_cascor_client.observability import record_unrecognized_frame